Reorder application of colors in RB_Jacobi
This order of coloring synchronizes well with the current default
interpolation for 1D problems.  In the current interpolation scheme, the even-
numbered points map injectively to the coarse-grid points and, in principle,
should be more accurate than the adjacent odd-numbered points whose values are
updated via interpolation.  Thus, in the red-black jacobi iteration, it makes
sense to update the odd-numbered points first, since the update of those values
depends entirely on the error in the even-numbered points.

I also added a disclaimer that the "block" aspect of the block Jacobi
has not been implemented yet.