Add color_flip option for smoother, change default color order
This option is used to determine whether the smoother traverses the list of
colors forwards or backwards.

Moreover, with the fix to the previous commit, I realize that it is actually
better to sweep in the other order.  There's something about sweeping color 0
first that reduces oscillatory errors that are not well-converged by multigrid.