Fix interpolation at boundaries and coarse grid setup in 1D mutligrid
The interpolation operator was not being defined properly at problem
boundaries.  This has been fixed.  Moreover, the coarse grid was supposed to
match up with the even-indexed fine grid points.  This was not the case before,
but it is the case now.

This fix should make it possible to use grid sizes that are not of the form
2^n+1.